5.3 -finalHand-In
Website is finished is scales and responds to desktop, laptop and
mobile. Mobile sensors include GPS and Camera implemented into  the
website. Sound is used as a backing track to give theme, video is on the
starting slide to set the scene of the website. Text is used though out
images are used for icons.

Website dev by Arapaoa Moffat (300317288)