11.5 - ifMarkerFoundWithInRadiusOfCircleAlert
If user marker position is within radius of disaster event marker than a
pop up box is show, if not it is hidden. This detects if a user is
within a disaster zone or outside it, if they are not within the radius
no alert is shown, if they are within the radius then a alert is shown.
Currently has a pop up box that is unwatned but working on removing it
now.